{{cookiecutter.project_slug}}/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.project_slug }}"
version = "0.0.1"
description = "{{ cookiecutter.project_description }}"
authors = ["{{ cookiecutter.project_owner }} <{{ cookiecutter.project_owner_email }}>"]

[tool.poetry.dependencies]
python = "^3.6"
psycopg2-binary = "2.8.3"
SQLAlchemy = "1.3.6"
Flask = "1.1.1"
marshmallow = "3.0.0"
gunicorn = "19.9.0"
Flask-Migrate = "^2.5.2"
Flask-SQLAlchemy = "2.4.0"
dynaconf = "2.2.2"
click = "^7.0"
alembic = "^1.3.1"

[tool.poetry.dev-dependencies]
ipython = "^7.8.0"
pytest = "^5.1.2"
pytest-cov = "^2.7.1"
pylint = "^2.4.3"
flake8 = "^3.7.8"
freezegun = "^0.3.12"
factory-boy = "^2.12.0"
pytest-env = "^0.6.2"

[tool.poetry.scripts]
initialize-db = "{{ cookiecutter.project_slug }}.scripts:initialize_db"
upgrade-db = "flask db upgrade"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
